Enhancement : Restructuring
