+++ data/got_memes/images/meme-got01.json
@@ -1,6 +1,6 @@
 {
     "mode": "0",
     "format": "1.1",
-    "description": "NewMeme",
-    "rating": 0
+    "description": "newMeme",
+    "rating": "0"
 }
