+++ index/memedb.json
@@ -11,7 +11,7 @@
     "Ellaria": [
         4
     ],
-    "NewMeme": [
+    "newMeme": [
         5
     ]
 }
