+++ services/logo.py
