+++ services/manipulation.py
